+++ frontend/src/components/auth/SignUp.tsx
@@ -1,23 +1,12 @@
-import Tippy from "@tippyjs/react";
 import axios from "axios";
-import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
-import { toast, ToastContainer } from "react-toastify";
+import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import "tippy.js/dist/tippy.css";
 import "../../css/auth/SignUp.css";
-import EmailIcon from "../utilities/icons/Email";
-import PasswordIcon from "../utilities/icons/Password";
-import Input from "../utilities/Input";
+import SignUpForm from "../utilities/SignUpForm";
 
 const SignUp = () => {
-  const {
-    register,
-    handleSubmit,
-    getValues,
-    formState: { errors },
-  } = useForm();
-
   const navigator = useNavigate();
 
   const handleFormSubmit = async (data: any) => {
@@ -40,42 +29,6 @@ const SignUp = () => {
       });
   };
 
-  const emailValidation = {
-    ...register("email", {
-      required: {
-        value: true,
-        message: "email is required",
-      },
-      pattern: {
-        value: /^\S+@\incubyte.co$/i,
-        message: "email must be an incubyte email",
-      },
-    }),
-  };
-
-  const passwordValidation = {
-    ...register("password", {
-      required: {
-        value: true,
-        message: "password is required",
-      },
-      pattern: {
-        value: /((?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[\W]).{6,20})/,
-        message: "password is not valid",
-      },
-    }),
-  };
-
-  const confirmPasswordValidation = {
-    ...register("confirmpassword", {
-      validate: {
-        confirmPasswordEqual: (value) =>
-          value === getValues().password ||
-          "Confirm Password must match with Password",
-      },
-    }),
-  };
-  const accountText = "Already have an account? ";
   return (
     <>
       <div className="SignUpContainer">
@@ -91,88 +44,10 @@ const SignUp = () => {
             <div className="SignUpHeader" data-testid="signupHeader">
               Sign Up
             </div>
-
-            <form onSubmit={handleSubmit(handleFormSubmit)} data-testid="form">
-              <label className="SignUpFormItems">
-                <Input
-                  icon={EmailIcon}
-                  dataTestId="signupEmail"
-                  placeholder="email"
-                  Id="email"
-                  type="text"
-                  showPasswordButton={false}
-                  validation={emailValidation}
-                />
-
-                <div
-                  data-testid="signupEmailError"
-                  id="signupEmailError"
-                  className="SignUpErrors"
-                >
-                  {errors.email ? <>{errors.email.message}</> : <></>}
-                </div>
-
-                <Input
-                  icon={PasswordIcon}
-                  dataTestId="signupPassword"
-                  placeholder="password"
-                  Id="password"
-                  showPasswordButton={true}
-                  validation={passwordValidation}
-                />
-
-                <Tippy content="password must contain 1 uppercase, 1 lowercase, 1 special character and 1 number">
-                  <div
-                    data-testid="signupPasswordError"
-                    className="SignUpErrors"
-                  >
-                    {errors.password ? <>{errors.password.message}</> : <></>}
-                  </div>
-                </Tippy>
-
-                <Input
-                  icon={PasswordIcon}
-                  dataTestId="signupConfirmPassword"
-                  placeholder="confirm password"
-                  Id="confirmpassword"
-                  showPasswordButton={true}
-                  validation={confirmPasswordValidation}
-                />
-                <div
-                  data-testid="signupConfirmPasswordError"
-                  className="SignUpErrors"
-                >
-                  {errors.confirmpassword ? (
-                    <>{errors.confirmpassword.message}</>
-                  ) : (
-                    <></>
-                  )}
-                </div>
-
-                <button
-                  className="SignUpSubmit"
-                  data-testid="signupButton"
-                  onClick={handleSubmit(handleFormSubmit)}
-                >
-                  Sign Up
-                </button>
-
-                <div data-testid="signupToast">
-                  <ToastContainer />
-                </div>
-
-                <label className="SignUpAlreadyAccount">
-                  {accountText}
-                  <a
-                    className="SignInLink"
-                    data-testid="signupAccountAlreadyExistsLink"
-                    href="/auth/signin"
-                  >
-                    Sign In
-                  </a>
-                </label>
-              </label>
-            </form>
+            <SignUpForm
+              dataTestId="signupForm"
+              handleFormSubmit={handleFormSubmit}
+            ></SignUpForm>
           </div>
         </div>
       </div>
